Update oneGroupGraph.java
In a square grid, two cells are connected if they share an edge and share the same value. Sharing an edge is up, down, left and right, but not diagonal. Given a square grid, determine the number of cells in each connected group of 1 values. There will be an array of queries, each one an integer. Create a return array of integers where each element is the number of groups in the matrix that have a size that matches the query.

 

For example, the input matrix is:

 

1 1 1 1 1 1
1 1 0 0 0 0
0 0 0 1 1 1
0 0 0 1 1 1
0 0 1 0 0 0
1 0 0 0 0 0
There are 4 connected groups of 1 values, 1 with 8 elements, 1 with 6, and 2 with 1 element. If there are q = 4 queries, queries = [6, 1, 8, 2], the results array is res = [1, 2, 1, 0].

 

Function Description 

Complete the function onesGroups in the editor below. The function must return an array of integers that represent the answers for each query.

 

onesGroups has the following parameter(s):

    grid[grid[0],...grid[n-1]]:  a square 2D array of integers that represent the grid

    queries[queries[0],...queries[q-1]]:  an array of integers that represent query values