Update connectedGroups.java
. Connected Groups
Relationships between people may be represented in a matrix as a series of binary digits. For example, the direct relationships for person 0 with persons 0 through 5 might be shown as 101100.  This means that person 0 knows persons 0, 2 and 3, the indices of each of the 1 values. A relationship is transitive. In other words, if person 0 knows person 2 and person 2 knows person 3, then person 0 knows person 3 through person 2.  A group is composed of all of the people who know one another, whether directly or transitively.

 

For example, consider the following relationships matrix:

 


	110
	110
	001
 

Persons 0 and 1 are connected, while person 2 is not. There are 2 groups.

 

Determine the number of groups represented in a matrix.

 

Note: The method signatures may vary a little depending on the requirements of your chosen language. For example, C language will have an argument that represents the number of rows and columns in the matrix.

 

Function Description 

Complete the function countGroups in the editor below. The function must return an integer representing the number of groups of people.

 

countGroups has the following parameter(s):

    related[related[0],...related[n-1]]:  an array of strings of binary digits related[i] that represent direct connections of people

 

Constraints

1 ≤ n ≤ 300
0 ≤ i < n
|related| = n
Each related[i] contains a binary string of n zeros and ones. related is a square matrix.